+++ include/mruby_integration/models/logging.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "mruby.h"
+
+extern RClass* Logging_class;
+
+void
+append_models_Logging(mrb_state*);