+++ src/mruby_integration/core/misc.cpp
@@ -12,16 +12,6 @@ mrb_set_config_flags(mrb_state* mrb, mrb_value) -> mrb_value
 }
 
 auto
-mrb_set_trace_log_level(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  mrb_int level;
-  mrb_get_args(mrb, "i", &level);
-
-  SetTraceLogLevel(level);
-  return mrb_nil_value();
-}
-
-auto
 mrb_open_url(mrb_state* mrb, mrb_value) -> mrb_value
 {
   char* url;
@@ -55,11 +45,6 @@ append_core_misc(mrb_state* mrb)
                     mrb->kernel_module,
                     "set_config_flags",
                     mrb_set_config_flags,
-                    MRB_ARGS_REQ(1));
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "set_trace_log_level",
-                    mrb_set_trace_log_level,
                     MRB_ARGS_REQ(1));
   mrb_define_method(
     mrb, mrb->kernel_module, "open_url", mrb_open_url, MRB_ARGS_REQ(1));